requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
kagglehub>=0.2.0
pyyaml>=6.0
tqdm>=4.65.0
requests>=2.31.0
matplotlib>=3.7.0